+++ repository/src/main/java/com/sj/repository/repository/SiteUserRepository.java
@@ -13,12 +13,12 @@ import com.sj.model.model.SiteUser;
 import com.sj.model.type.ActivateEnum;
 
 public interface SiteUserRepository extends
-		PagingAndSortingRepository<SiteUser, Integer> {
+		PagingAndSortingRepository<SiteUser, Long> {
 	public SiteUser findByNameAndEnabled(String name, ActivateEnum activate);
 
 	public Page<SiteUser> findAll(Pageable pageable);
 	
 	@Modifying	
 	@Query("update SiteUser u set u.enabled=:enabled,u.authenticatedTime=:time where u.id=:id")
-	public void updateEnabled(@Param("enabled")ActivateEnum state,@Param("time")Calendar time,@Param("id")int id);
+	public void updateEnabled(@Param("enabled")ActivateEnum state,@Param("time")Calendar time,@Param("id")Long id);
 }
